Fix annoying configuration
** Karabiner **
Karabiner had the default keyboard listed as ansi instead of ISO....
which is beside the point as it doesn't seem to be listed in the damn
configuration file....

Swapped that around now so the swapping of the ~ with the § is no longer
needed. Why does it default to ansi?????

Added a mapping of CAPS_LOCK to the ESCAPE key so that it is easier to
use that in conjunction with neovim. Note to self as well the left
control key seems super far away from the rest of the keys. Keep and eye
on this as it could be really bad......

** Neovim **
Neovim's ansible role didn't correctly create the `carefulnow/plugins`
directory after the update to the configuration file path.

Remove the <leader>o shortcut as it's the exact same as <C-w><C-w>

Increase the scrolloff to 20 characters

Fix shortcut to run test method with jdtls to be <leader>dm instead of
<leader>dn